Refactor into alg package.
