+++ src/main/java/com/graphhopper/sna/alg/GraphSearchAlgorithmStartNode.java
@@ -2,7 +2,7 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.graphhopper.sna.centrality;
+package com.graphhopper.sna.alg;
 
 import org.jgrapht.Graph;
